Fixed some of the physics
The intersection function only works when the player is not on the surface of the polygon. The player is getting moved onto the surface of the polygon due to how the physics equations work. Another check needs to be added to check if the player is on the surface of the polygon.