Working on collision physics
There is a function that is a work in progress that checks if a 3D point is on a 3D triangle with 3 vertices. This is to fix the issue where the player goes through the mesh. Currently, it just checks if the point is perpendicular to the surface normal, and next, I will be adding a check to see if it is in bounds of the edges.