An Optimized Renderer And Work in Progress Physics
This file contains the slightly more optimized renderer (the difference is the vertices are being rotated than assigned to the polygons so each vertice isn't being rotated multiple times giving about 2-5 fps) and the work in progress physics. The physics are currently only working on flat ground but they will be improving. In this version the map is hard coded.